requirements.txt
pytesseract~=0.3.10
Pillow~=9.3.0
Levenshtein~=0.21.1